[frontend] initial commit for watchlist API endpoint
