+++ apps/frontend/src/components/modules/UserProfile/UserProfile.jsx
@@ -1,36 +1,26 @@
 import { Center, Spinner } from "@chakra-ui/react";
 import axios from "axios";
-import { useMemo } from "react";
 import ErrorLayout from "src/components/common/layouts/ErrorLayout";
 import useSWR from "swr";
 import UserDetailsGrid from "./UserDetailsGrid";
 import UserItemsGrid from "./UserItemsGrid";
 import WatchlistGrid from "./WatchlistGrid";
-import stubItems from "src/lib/stubData";
-import { useFirestoreUserData } from "src/lib/hooks";
+// import stubItems from "src/lib/stubData";
+import { useFirestoreUserData, useStoredAddress } from "src/lib/hooks";
 
 const UserProfile = () => {
   const { user, userData, error } = useFirestoreUserData();
-
-  const storedAddress = useMemo(
-    () => (userData ? userData.get("walletAddress") : null),
-    [userData]
-  );
-
-  const watchlistArray = useMemo(
-    () => (userData ? userData.get("watchlist") : null),
-    [userData]
-  );
+  const storedAddress = useStoredAddress(userData);
 
   const { data: userItems } = useSWR(
     storedAddress ? "/api/listings/user/" + storedAddress : null,
     (url) => axios.get(url).then((res) => res.data.msg)
   );
 
-  /* const { data: watchlist } = useSWR(
-    ["/api/listings/watchlist/", watchlistArray],
-    (url, array) =>
-  ); */
+  const { data: watchlist } = useSWR(
+    user ? "/api/listings/user/watchlist/" + user.uid : null,
+    (url) => axios.get(url).then((res) => res.data.msg)
+  );
 
   if (error) {
     return <ErrorLayout />;
@@ -43,7 +33,11 @@ const UserProfile = () => {
       ) : (
         <Spinner size="xl" color="gray" />
       )}
-      <WatchlistGrid watchlist={stubItems.userListings} />
+      {watchlist ? (
+        <WatchlistGrid watchlist={watchlist} />
+      ) : (
+        <Spinner size="xl" color="gray" />
+      )}
       {userItems ? (
         <UserItemsGrid items={userItems} />
       ) : (
